configuried env vars
