+++ app.js
@@ -2,7 +2,9 @@ const express = require('express')
 const logger = require('morgan')
 const cors = require('cors')
 
-const contactsRouter = require('./routes/api/controllers')
+require("dotenv").config()
+
+const contactsRouter = require('./routes/api/contacts-routes')
 
 const app = express()
