+++ package.json
@@ -11,8 +11,10 @@
   "dependencies": {
     "cors": "2.8.5",
     "cross-env": "7.0.3",
+    "dotenv": "^16.0.3",
     "express": "4.17.1",
     "joi": "^17.9.1",
+    "mongoose": "^7.0.3",
     "morgan": "1.10.0",
     "nanoid": "^3.3.4"
   },
