--- routes/api/controllers.js
@@ -1,133 +1,0 @@
-const express = require("express");
-const Joi = require("joi");
-
-const contacts = require("../../models/contacts");
-
-const { HttpError } = require("../../helpers");
-
-const router = express.Router();
-
-const addSchema = Joi.object({
-  name: Joi.string().min(3).required().messages({
-    "any.required": "Name must not be empty",
-    "string.empty": "Name must not be empty",
-    "string.min": "Name must be at least 3 characters",
-  }),
-  email: Joi.string()
-    .email()
-    .pattern(/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/)
-    .required()
-    .messages({
-      "any.required": "Email must not be empty",
-      "string.empty": "Email must not be empty",
-      "string.pattern.base": "Email does not have a right format",
-    }),
-  phone: Joi.string()
-    .required()
-    .pattern(/^\(\d{3}\)\s\d{3}-\d{4}$/)
-    .messages({
-      "any.required": "phone must not be empty",
-      "string.empty": "phone must not be empty",
-      "string.pattern.base":
-        "Phone number must be in the format (XXX) XXX-XXXX",
-    }),
-});
-
-const updateSchema = Joi.object({
-  name: Joi.string().min(3).messages({
-    "string.min": "Name must be at least 3 characters",
-  }),
-  email: Joi.string()
-    .pattern(/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/)
-    .messages({
-      "string.email": "Email mast be valid",
-      "string.pattern.base": "Email does not have a right format",
-    }),
-  phone: Joi.string()
-    .pattern(/^\(\d{3}\)\s\d{3}-\d{4}$/)
-    .messages({
-      "string.empty": "phone must not be empty",
-      "string.pattern.base":
-        "Phone number must be in the format (XXX) XXX-XXXX",
-    }),
-}).or("name", "email", "phone");
-
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (err) {
-    next(err);
-  }
-});
-
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const {contactId: id } = req.params;
-    const result = await contacts.getContactById(id);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.post("/", async (req, res, next) => {
-  try {
-    if (Object.keys(req.body).length < 3) {
-      throw HttpError(400, "missing required name field");
-    }
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  } catch (err) {
-    next(err);
-  }
-});
-
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId: id } = req.params;
-    const result = await contacts.removeContact(id);
-    if (!result) {
-      throw HttpError(404);
-    }
-    res.json({
-      message: "contact deleted",
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    if (!Object.keys(req.body).length) {
-      throw HttpError(400);
-    }
-    const { error } = updateSchema.validate(req.body);
-    console.log("error", error);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    const { contactId: id } = req.params;
-    const result = await contacts.updateContact(id, req.body);
-    console.log("body", req.body);
-    console.log("result", result);
-    if (!result) {
-      throw HttpError(404);
-    }
-
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-module.exports = router;
