+++ utils/ctrlWrapper.js
@@ -1,0 +1,13 @@
+const crtlWrapper = ctrl  => {
+    const func = async (req, res, next) => {
+        try {
+            await ctrl(req, res, next);
+         }
+        catch (err) {
+            next(err)
+        }
+    }
+    return func;
+}
+
+module.exports = crtlWrapper;
