+++ utils/validateBody.js
@@ -1,0 +1,14 @@
+const HttpError=require('../helpers/HttpError')
+
+const validateBody = schema => {
+    const func = async (req, res, next) => {
+        const { error } = schema.validate(req.body);
+        if (error) {
+            return next(HttpError(400, error.message));
+        }
+        next();
+    }
+    return func;
+}
+
+module.exports = validateBody;
